+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## [0.1.2] - 2020-03-19
+
+### Added
+- TMC-Readme.txt in the TMC workspace
+- Marketplace icon
+
+## Updated
+- Fix: Fix data validation issue with very old courses in the HY organization
+
 ## [0.1.1] - 2020-03-18
 
 ### Updated